Split old ChaseTarget into ChaseTarget and VirusTargetPicker to improve capatibility. Added "Bullet" tag.
