+++ Assets/Scripts/Ye/VirusTargetPicker.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 81bfaeec9ab62d140a4c0d8dc666d5b2
+timeCreated: 1488371797
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
